+++ package.json
@@ -27,7 +27,7 @@
     "build"
   ],
   "scripts": {
-    "build": "tsc --build tsconfig.json && node automation-post-compile.js",
+    "build": "tsc --build tsconfig.json",
     "clean": "node automation-clean.js",
     "format": "sort-package-json && prettier . --write && pnpm eslint --fix . --ext .js,.ts",
     "lint": "sort-package-json --check && prettier . --check && pnpm eslint . --ext .js,.ts",
