+++ src/internal/domain/Selector.ts
