+++ test/acceptance/IgnoredSelectorsTest.spec.ts
