+++ test/acceptance/LeaveKeyframePercentagesTest.spec.ts
