+++ test/acceptance/ReplacementPrefixing.spec.ts
