+++ test/acceptance/fixtures/leave-body-expected.css
