+++ test/acceptance/fixtures/leave-expected.css
