+++ test/acceptance/fixtures/standard-attribute-selectors-expected.css
