+++ test/acceptance/fixtures/standard-universal-selectors-expected.css
