+++ test/acceptance/fixtures/standard-universal-selectors-raw.css
