+++ test/acceptance/support/PluginBootstrap.ts
@@ -1,9 +1,9 @@
 import PostCSS from "postcss";
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
-import PrefixWrap from "../../index";
+import PrefixWrap from "../../../src";
 
-import { PostCSSPrefixWrapOptions } from "PostCSSPrefixWrap";
+import { PostCSSPrefixWrapOptions } from "../../../src/plugin/PostCSSPrefixWrap";
 import Processor from "postcss/lib/processor";
 
 export const postCSSWithPlugin = (
