+++ test/unit/plugin/indexWrap.spec.ts
@@ -1,4 +1,4 @@
-import indexWrap from "./indexWrap";
+import indexWrap from "../../../src/plugin/indexWrap";
 import * as postcss8 from "postcss";
 import * as postcss7 from "postcss7";
